Extract ReleasePatchVersion to its own class and ReleaserUtils module
